# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Utility Management System API",
        "description": "API documentation for the Laravel Utility Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a list of rooms",
                "description": "Returns a paginated list of all rooms.",
                "operationId": "7f4754b7db3173fe97eae61d624a10f4",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rooms retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RoomResource"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 15
                                                        },
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 4
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rooms not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a single room",
                "description": "Returns the details of a specific room by its UUID.",
                "operationId": "5b14778d4ad1f8dd7d7b27f77cc6d099",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the room",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update an existing room",
                "description": "Updates the details of an existing room.",
                "operationId": "98cd7940d692e048c3bc38cd5654eaea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the room to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roomNo",
                                    "floor",
                                    "dimension",
                                    "noOfBedRoom",
                                    "status",
                                    "sellingPrice",
                                    "maxNoOfPeople"
                                ],
                                "properties": {
                                    "roomNo": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "floor": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dimension": {
                                        "type": "string",
                                        "example": "12x12 sqft"
                                    },
                                    "noOfBedRoom": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "Available",
                                            "Rented",
                                            "Purchased",
                                            "In Maintenance"
                                        ],
                                        "example": "Available"
                                    },
                                    "sellingPrice": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150000
                                    },
                                    "maxNoOfPeople": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A cozy room with a nice view.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Room not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoomResource": {
                "title": "Room Resource",
                "description": "Room model representation",
                "properties": {
                    "id": {
                        "description": "Room's unique identifier",
                        "type": "string",
                        "format": "uuid"
                    },
                    "roomNo": {
                        "description": "Room number",
                        "type": "integer",
                        "example": 101
                    },
                    "floor": {
                        "description": "Floor number",
                        "type": "integer",
                        "example": 1
                    },
                    "dimension": {
                        "description": "Room dimensions",
                        "type": "string",
                        "example": "12x12 sqft"
                    },
                    "noOfBedRoom": {
                        "description": "Number of bedrooms",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "Current status of the room",
                        "type": "string",
                        "enum": [
                            "Available",
                            "Rented",
                            "Purchased",
                            "In Maintenance"
                        ],
                        "example": "Available"
                    },
                    "sellingPrice": {
                        "description": "Selling price of the room",
                        "type": "number",
                        "format": "float",
                        "example": 150000
                    },
                    "maxNoOfPeople": {
                        "description": "Maximum number of occupants",
                        "type": "integer",
                        "example": 2
                    },
                    "description": {
                        "description": "A brief description of the room",
                        "type": "string",
                        "example": "A cozy room with a nice view.",
                        "nullable": true
                    },
                    "createdAt": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}